requirements.txt
Flask
Flask-CORS
PyMuPDF
requests
google-generativeai
reportlab
python-dotenv
matplotlib


